+++ src/preview.ts
@@ -189,9 +189,9 @@ export default class URDFPreview
             this._webview.webview.postMessage({ 
                 command: 'colors', 
                 cameraRadius: config.get("CameraDistanceToRobot", "1.0"),
-                cameraAlpha: config.get("CameraAlpha", -Math.PI / 3),
-                cameraBeta: config.get("CameraBeta", 5 * Math.PI / 12),
-                cameraRadiusNew: config.get("CameraRadius", 1),
+                defaultCameraAlpha: config.get("CameraAlpha", -Math.PI / 3),
+                defaultCameraBeta: config.get("CameraBeta", 5 * Math.PI / 12),
+                defaultCameraRadius: config.get("CameraRadius", 1),
                 backgroundColor: config.get("BackgroundColor", "#000000"),
                 gridLineColor: config.get("GridMinorColor", "#00FF00"),
                 gridMainColor: config.get("GridMainColor", "#001100"),